Add files via upload
Event.java = An Event object that holds an ID number, whether it's a hang up Event or a dial-in Event, and the arrival time of the caller
Poisson.java = A Poisson distribution class that my professor told us we could use. I did not code this class
PriorityQueue.java = A priority queue class that prioritizing who calls first 
WaitingQueue.java = A waiting queue class of callers waiting to get into the priority queue
Simulation.java = The main method of the project that actually runs the simulation. This is the only class in the project that is coded 100% by me.